refactor(api): Moves api-hooks from api/func to api
